Refactor Exercise management and enhance UI with tests (#7)
* Large Views Decomposition + Tests

* delete confirmation added where applicable

* ExerciseSet and WorkoutRepository tests

* ExerciseSet and WorkoutRepository tests

* Drop files from .gitignore

* Tests fix

* Icons and adaptive glass style

* UI adjustments and optimizations

* README update

* Main (#5)

* Add screenshots section to README

Added screenshots to the README file.

* Screenshots alignment in README

* Icons

* duplicates

* separating ExerciseList and ExercisePicker logic

* ExercisePickerView in progress

* - Centralized Localization
- OrderedSet test
- ExercisePickerView partial

* - ExercisePickerRowView ready

* SwiftLint added + formatting

* multiple choice of exercise for Workout + tests

* ExerciseSetListView section header fix

* formatting

* Progress functionality implementation

* Progress implementation Repository + ViewModel

* - Localization fix
- Add Button moved
- Size of Rows in Views is smaller
- Paddings are more compact
- Font sizes a bit smaller

* Add button on WorkoutDetailView moved from bottom to top toolbar

* Set's order logic is changed

* - Exercise Muscle Group
- Warm up sets

* Sets View fix

* Muscle group duplicate removed

* Task lifecycle added

* CoreDataWorkoutRepository Optimization

* Updated Errors + AddExerciseToWorkoutUseCase

* Updated Errors + AddExerciseToWorkoutUseCase

* DeleteWorkoutUseCase + ManageWorkoutTagsUseCase

* Preparations for ExerciseProgressView

* ExerciseProgressView (Part 1, Period Picker)

* ExerciseProgressView (Part 1, Period Picker)

* ExerciseProgressVolumeChartView

* - Documentation updated
- mistype Appear
- selectedPeriod update logic changed
- WorkoutRepository devided on 3 protocols

* Progress agregation logic moved to UseCase

* Moved Concreate instantation our of AppDependecies

* ProgressView

* small fixes

* ManageExerciseUseCase and FetchExercisesUseCase created

* UseCase in ExercisePickerViewModel

* Project update

* CloudKit support

* import/export functionality

* restore functionality removed

* project settings

---------

+++ Liftlog/Core/Repositories/ExerciseRepository/CoreDataExerciseRepository.swift
@@ -59,9 +59,9 @@ final class CoreDataExerciseRepository: ExerciseRepositoryProtocol {
             let exercise = Exercise(context: self.context)
             exercise.id = UUID()
             exercise.name = name
-            exercise.type = Int16(type.rawValue)
+            exercise.type = Int32(type.rawValue)
             exercise.exerciseDescription = description
-            exercise.muscleGroup = Int16(muscleGroup?.rawValue ?? -1)
+            exercise.muscleGroup = Int32(muscleGroup?.rawValue ?? -1)
 
             try self.context.saveOrThrow()
 
@@ -74,9 +74,9 @@ final class CoreDataExerciseRepository: ExerciseRepositoryProtocol {
             let exercise = try self.fetchExercise(model.id)
 
             exercise.name = model.name
-            exercise.type = Int16(model.type.rawValue)
+            exercise.type = Int32(model.type.rawValue)
             exercise.exerciseDescription = model.description
-            exercise.muscleGroup = Int16(model.muscleGroup?.rawValue ?? -1)
+            exercise.muscleGroup = Int32(model.muscleGroup?.rawValue ?? -1)
 
             try self.context.saveOrThrow()
         }
